# Use byte slicing to reduce cost of multiple message decoding