--- a/fuzz_test.go
+++ b/fuzz_test.go
@@ -7,11 +7,29 @@ import (
 	"testing"
 )
 
-func FuzzJSONDecode(f *testing.F) {
+func FuzzJSONDecodeSingle(f *testing.F) {
 	f.Add([]byte(`{"id": 1, "method": "", "params": {}}`))
 	f.Fuzz(func(t *testing.T, b []byte) {
 		decoder := GetCommandDecoder(TypeJSON, b)
 		_, err := decoder.Decode()
+		if err != nil {
+			t.Skip()
+		}
+		PutCommandDecoder(TypeJSON, decoder)
+	})
+}
+
+func FuzzJSONDecodeMultiple(f *testing.F) {
+	f.Add([]byte(`{"id": 1, "method": "", "params": {}}
+{"id": 2, "method": "", "params": {}}
+`))
+	f.Fuzz(func(t *testing.T, b []byte) {
+		decoder := GetCommandDecoder(TypeJSON, b)
+		_, err := decoder.Decode()
+		if err != nil {
+			t.Skip()
+		}
+		_, err = decoder.Decode()
 		if err != nil {
 			t.Skip()
 		}