Instead of json.Decoder we can just slice initial data since we have a strict format for expected JSON input:

```
{}
{}
{}
```

I.e. individual JSON objects separated by exactly one new line (with optional new line at the end).

Benchmarks:

```
name                                   old time/op    new time/op    delta
ReplyJSONUnmarshalMultiple-12            4.18µs ± 1%    0.83µs ± 1%  -80.21%  (p=0.000 n=9+8)
ReplyJSONUnmarshalMultipleParallel-12    3.32µs ± 4%    0.22µs ± 9%  -93.32%  (p=0.000 n=10+10)

name                                   old alloc/op   new alloc/op   delta
ReplyJSONUnmarshalMultiple-12            33.3kB ± 0%     0.5kB ± 0%  -98.61%  (p=0.000 n=10+10)
ReplyJSONUnmarshalMultipleParallel-12    33.3kB ± 0%     0.5kB ± 0%  -98.61%  (p=0.000 n=10+10)

name                                   old allocs/op  new allocs/op  delta
ReplyJSONUnmarshalMultiple-12              8.00 ± 0%      6.00 ± 0%  -25.00%  (p=0.000 n=10+10)
ReplyJSONUnmarshalMultipleParallel-12      8.00 ± 0%      6.00 ± 0%  -25.00%  (p=0.000 n=10+10)
```